requirements.txt
jax
dm-haiku
optax
tensorflow_datasets
